# backend/rest/swagger.json
{
    "swagger": "2.0",
    "info": {
      "description": "Marc Recipes API",
      "version": "1.0.0",
      "title": "Marc Recipes API",
      "contact": {
        "name": "RealWorld",
        "url": "https://realworld.io"
      },
      "license": {
        "name": "MIT License",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "localhost:3000",
    "basePath": "/api",
    "schemes": [
      "http"
    ],
    "produces": [
      "application/json"
    ],
    "consumes": [
      "application/json"
    ],
    "securityDefinitions": {
      "Token": {
        "description": "For accessing the protected API resources, you must have received a a valid JWT token after registering or logging in. This JWT token must then be used for all protected resources by passing it in via the 'Authorization' header.\n\nA JWT token is generated by the API by either registering via /users or logging in via /users/login.\n\nThe following format must be in the 'Authorization' header :\n\n    Token: xxxxxx.yyyyyyy.zzzzzz\n    \n",
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    },
    "paths": {
...
      "/foods/feed": {
        "get": {
          "summary": "Get recent recipes from users you follow",
          "description": "Get most recent recipes from users you follow. Use query parameters to limit. Auth is required",
          "tags": [
            "Recipes"
          ],
          "security": [
            {
              "Token": []
            }
          ],
          "operationId": "GetRecipesFeed",
          "parameters": [
            {
              "name": "limit",
              "in": "query",
              "description": "Limit number of recipes returned (default is 20)",
              "required": false,
              "default": 20,
              "type": "integer"
            },
            {
              "name": "offset",
              "in": "query",
              "description": "Offset/skip number of recipes (default is 0)",
              "required": false,
              "default": 0,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/MultipleFoodsResponse"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        }
      },
      "/foods": {
        "get": {
          "summary": "Get recent recipes globally",
          "description": "Get most recent recipes globally. Use query parameters to filter results. Auth is optional",
          "tags": [
            "Recipes"
          ],
          "operationId": "GetRecipes",
          "parameters": [
            {
              "name": "tag",
              "in": "query",
              "description": "Filter by tag",
              "required": false,
              "type": "string"
            },
            {
              "name": "author",
              "in": "query",
              "description": "Filter by author (username)",
              "required": false,
              "type": "string"
            },
            {
              "name": "favorited",
              "in": "query",
              "description": "Filter by favorites of a user (username)",
              "required": false,
              "type": "string"
            },
            {
              "name": "limit",
              "in": "query",
              "description": "Limit number of recipes returned (default is 20)",
              "required": false,
              "default": 20,
              "type": "integer"
            },
            {
              "name": "offset",
              "in": "query",
              "description": "Offset/skip number of recipes (default is 0)",
              "required": false,
              "default": 0,
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/MultipleFoodsResponse"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        },
        "post": {
          "summary": "Create a recipe",
          "description": "Create an recipe. Auth is required",
          "tags": [
            "Recipes"
          ],
          "security": [
            {
              "Token": []
            }
          ],
          "operationId": "CreateRecipe",
          "parameters": [
            {
              "name": "food",
              "in": "body",
              "required": true,
              "description": "Food to create",
              "schema": {
                "$ref": "#/definitions/NewFoodRequest"
              }
            }
          ],
          "responses": {
            "201": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/SingleFoodResponse"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        }
      },
      "/foods/{slug}": {
        "get": {
          "summary": "Get a recipe",
          "description": "Get a recipe. Auth not required",
          "tags": [
            "Recipes"
          ],
          "operationId": "GetFood",
          "parameters": [
            {
              "name": "slug",
              "in": "path",
              "required": true,
              "description": "Slug of the recipe to get",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/SingleFoodResponse"
              }
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        },
        "put": {
          "summary": "Update a recipe",
          "description": "Update a recipe. Auth is required",
          "tags": [
            "Recipes"
          ],
          "security": [
            {
              "Token": []
            }
          ],
          "operationId": "UpdateFood",
          "parameters": [
            {
              "name": "slug",
              "in": "path",
              "required": true,
              "description": "Slug of the recipe to update",
              "type": "string"
            },
            {
              "name": "food",
              "in": "body",
              "required": true,
              "description": "Recipe to update",
              "schema": {
                "$ref": "#/definitions/UpdateFoodRequest"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/SingleFoodResponse"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        },
        "delete": {
          "summary": "Delete a recipe",
          "description": "Delete a recipe. Auth is required",
          "tags": [
            "Recipes"
          ],
          "security": [
            {
              "Token": []
            }
          ],
          "operationId": "DeleteFood",
          "parameters": [
            {
              "name": "slug",
              "in": "path",
              "required": true,
              "description": "Slug of the recipe to delete",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        }
      },
      "/foods/{slug}/comments": {
        "get": {
          "summary": "Get comments for a recipe",
          "description": "Get the comments for a recipe. Auth is optional",
          "tags": [
            "Comments"
          ],
          "operationId": "GetFoodComments",
          "parameters": [
            {
              "name": "slug",
              "in": "path",
              "required": true,
              "description": "Slug of the recipe that you want to get comments for",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/MultipleCommentsResponse"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        },
        "post": {
          "summary": "Create a comment for a recipe",
          "description": "Create a comment for a recipe. Auth is required",
          "tags": [
            "Comments"
          ],
          "security": [
            {
              "Token": []
            }
          ],
          "operationId": "CreateFoodComment",
          "parameters": [
            {
              "name": "slug",
              "in": "path",
              "required": true,
              "description": "Slug of the recipe that you want to create a comment for",
              "type": "string"
            },
            {
              "name": "comment",
              "in": "body",
              "required": true,
              "description": "Comment you want to create",
              "schema": {
                "$ref": "#/definitions/NewCommentRequest"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/SingleCommentResponse"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        }
      },
      "/foods/{slug}/comments/{id}": {
        "delete": {
          "summary": "Delete a comment for a recipe",
          "description": "Delete a comment for a recipe. Auth is required",
          "tags": [
            "Comments"
          ],
          "security": [
            {
              "Token": []
            }
          ],
          "operationId": "DeletFoodComment",
          "parameters": [
            {
              "name": "slug",
              "in": "path",
              "required": true,
              "description": "Slug of the recipe that you want to delete a comment for",
              "type": "string"
            },
            {
              "name": "id",
              "in": "path",
              "required": true,
              "description": "ID of the comment you want to delete",
              "type": "integer"
            }
          ],
          "responses": {
            "200": {
              "description": "OK"
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        }
      },
      "/foods/{slug}/favorite": {
        "post": {
          "summary": "Favorite a recipe",
          "description": "Favorite a recipe. Auth is required",
          "tags": [
            "Favorites"
          ],
          "security": [
            {
              "Token": []
            }
          ],
          "operationId": "CreateFoodFavorite",
          "parameters": [
            {
              "name": "slug",
              "in": "path",
              "required": true,
              "description": "Slug of the recipe that you want to favorite",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/SingleFoodResponse"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        },
        "delete": {
          "summary": "Unfavorite a recipe",
          "description": "Unfavorite a recipe. Auth is required",
          "tags": [
            "Favorites"
          ],
          "security": [
            {
              "Token": []
            }
          ],
          "operationId": "DeleteFoodFavorite",
          "parameters": [
            {
              "name": "slug",
              "in": "path",
              "required": true,
              "description": "Slug of the recipe that you want to unfavorite",
              "type": "string"
            }
          ],
          "responses": {
            "200": {
              "description": "OK",
              "schema": {
                "$ref": "#/definitions/SingleFoodResponse"
              }
            },
            "401": {
              "description": "Unauthorized"
            },
            "422": {
              "description": "Unexpected error",
              "schema": {
                "$ref": "#/definitions/GenericErrorModel"
              }
            }
          }
        }
      },
...
    },
    "definitions": {
...
      "Profile": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "image": {
            "type": "string"
          },
          "following": {
            "type": "boolean"
          }
        },
        "required": [
          "username",
          "bio",
          "image",
          "following"
        ]
      },
      "Food": {
        "type": "object",
        "properties": {
          "slug": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "tagList": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "favorited": {
            "type": "boolean"
          },
          "favoritesCount": {
            "type": "integer"
          },
          "author": {
            "$ref": "#/definitions/Profile"
          }
        },
        "required": [
          "slug",
          "title",
          "description",
          "body",
          "tagList",
          "createdAt",
          "updatedAt",
          "favorited",
          "favoritesCount",
          "author"
        ]
      },
      "SingleFoodResponse": {
        "type": "object",
        "properties": {
          "food": {
            "$ref": "#/definitions/Food"
          }
        },
        "required": [
          "food"
        ]
      },
      "MultipleFoodsResponse": {
        "type": "object",
        "properties": {
          "foods": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Food"
            }
          },
          "foodsCount": {
            "type": "integer"
          }
        },
        "required": [
          "foods",
          "foodsCount"
        ]
      },
      "NewFood": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "tagList": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "title",
          "description",
          "body"
        ]
      },
      "NewFoodRequest": {
        "type": "object",
        "properties": {
          "food": {
            "$ref": "#/definitions/NewFood"
          }
        },
        "required": [
          "food"
        ]
      },
      "UpdateFood": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "body": {
            "type": "string"
          }
        }
      },
      "UpdateFoodRequest": {
        "type": "object",
        "properties": {
          "food": {
            "$ref": "#/definitions/UpdateFood"
          }
        },
        "required": [
          "food"
        ]
      },
      "Comment": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "body": {
            "type": "string"
          },
          "author": {
            "$ref": "#/definitions/Profile"
          }
        },
        "required": [
          "id",
          "createdAt",
          "updatedAt",
          "body",
          "author"
        ]
      },
      "SingleCommentResponse": {
        "type": "object",
        "properties": {
          "comment": {
            "$ref": "#/definitions/Comment"
          }
        },
        "required": [
          "comment"
        ]
      },
      "MultipleCommentsResponse": {
        "type": "object",
        "properties": {
          "comments": {
            "type": "array",
            "items": {
              "$ref": "#/definitions/Comment"
            }
          }
        },
        "required": [
          "comments"
        ]
      },
      "NewComment": {
        "type": "object",
        "properties": {
          "body": {
            "type": "string"
          }
        },
        "required": [
          "body"
        ]
      },
      "NewCommentRequest": {
        "type": "object",
        "properties": {
          "comment": {
            "$ref": "#/definitions/NewComment"
          }
        },
        "required": [
          "comment"
        ]
      },
...
      "GenericErrorModel": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "object",
            "properties": {
              "body": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "required": [
              "body"
            ]
          }
        },
        "required": [
          "errors"
        ]
      }
    }
  }